Implement tree pool (#29)
* implemented an option to select random points to compose the inverse generated pool

* fixed bug not adding two lists

* Implements first version of the TreePool

The _ask method is stucked selecting the same branch. Will be fixed in future commits

* changed some tests to compare approximately

 Using int to avoid decimal number caused an error where 3.999...!=4

* added TreePool to the import

* implemented tests for TreePool

* adapted a test to compare approximately

Using int to convert the prediction output cause the test to fail when asserting 3.9999...==4. Using pytest.approx() to solve that

* Solved the issue of ask() always selecting the same branch

Did that by randomly sampling when it selects the same branch. Also implemented a leaf deletion for BO

* bumped version

* Removed python 3.8 from the pipeline. OrderedDict can only be used as type hinting after python 3.9